# contracts/openapi.yaml
openapi: 3.1.1
info:
  title: ACCP Collaboration API
  version: 0.4.0
  description: ACCP M1–M4 runtime API. Human ownership, delegated execution, verified artifacts, independent approvals, MCP Gateway and React console. The existing 0.2 Adapter handshake remains compatible. Real AI-client joint acceptance is deferred by the project owner; see docs/m3-m4-platform.md.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: ACCP maintainers
    url: https://github.com/JavaWeh/ACCP
servers:
  - url: https://accp.example/api/v1
    description: Reserved example host; not a deployment
...
security:
  - HumanBearer: []
  - AgentBearer: []
paths:
...
  /projects/{id}/repositories:
    get:
      operationId: listRepositories
      tags:
        - Tasks
      summary: List configured repositories
      description: Read project repository configuration; first provider planned is GitHub.
      x-accp-actors:
        - HUMAN
      security:
        - HumanBearer: []
      parameters:
        - name: id
          in: path
          required: true
          description: Opaque resource identifier; scope is verified server-side.
          schema:
            type: string
            minLength: 1
        - name: cursor
          in: query
          required: false
          description: Opaque continuation cursor.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Maximum page size.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: search
          in: query
          required: false
          description: Optional server filter; changing filters invalidates the cursor. Unfiltered requests preserve legacy ID ordering.
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: Optional server filter; changing filters invalidates the cursor. Unfiltered requests preserve legacy ID ordering.
          schema:
            type: string
        - name: owner
          in: query
          required: false
          description: Optional server filter; changing filters invalidates the cursor. Unfiltered requests preserve legacy ID ordering.
          schema:
            type: string
        - name: sort
          in: query
          required: false
          description: Optional server filter; changing filters invalidates the cursor. Unfiltered requests preserve legacy ID ordering.
          schema:
            type: string
            enum:
              - id
              - title
              - updated
              - -updated
      responses:
        "200":
          description: Successful operation. Human ownership and resource scopes are validated by the server.
          headers:
            ETag:
              description: Quoted resource version when returning a versioned entity.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: ./schemas/domain.schema.json#/$defs/RepositoryPage
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
      x-accp-milestone: M1
    post:
      operationId: registerProjectRepository
      tags:
        - Identity
      summary: register Project Repository
      description: Human administration. Organization operations require an active project administrator in the same enterprise. Project writes require ADMIN in that project. Archive requires all tasks and operations to be settled and revokes remaining Sessions. No membership is inferred from identity registration.
      x-accp-milestone: M4
      x-accp-actors:
        - HUMAN
      security:
        - HumanBearer: []
      parameters:
        - name: id
          in: path
          required: true
          description: Project ID; requires current membership.
          schema:
            type: string
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: Successful administration operation.
          content:
            application/json:
              schema:
                $ref: ./schemas/domain.schema.json#/$defs/Repository
          headers:
            ETag:
              description: Quoted resource version when applicable.
              schema:
                type: string
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "412":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "428":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas/management.schema.json#/$defs/RegisterRepository
  /projects/{id}/tasks:
    post:
      operationId: createTask
      tags:
        - Tasks
      summary: Create a human-owned task
      description: Creates DRAFT. Owner must be an active human member; repository and Context versions must belong to this project.
      x-accp-actors:
        - HUMAN
      security:
        - HumanBearer: []
      parameters:
        - name: id
          in: path
          required: true
          description: Opaque resource identifier; scope is verified server-side.
          schema:
            type: string
            minLength: 1
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: Successful operation. Human ownership and resource scopes are validated by the server.
          headers:
            ETag:
              description: Quoted resource version when returning a versioned entity.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: ./schemas/domain.schema.json#/$defs/Task
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas/domain.schema.json#/$defs/CreateTask
      x-accp-milestone: M1
    get:
      operationId: listTasks
      tags:
        - Tasks
      summary: List project tasks
      description: Read authorized project tasks. M2 also allows a scoped Agent Session in the same project.
      x-accp-actors:
        - HUMAN
        - AGENT
      security:
        - HumanBearer: []
        - AgentBearer: []
      parameters:
        - &a6
          name: id
          in: path
          required: true
          description: Opaque resource identifier; scope is verified server-side.
          schema:
            type: string
            minLength: 1
        - name: cursor
          in: query
          required: false
          description: Opaque continuation cursor.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Maximum page size.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: search
          in: query
          required: false
          description: Optional server filter; changing filters invalidates the cursor. Unfiltered requests preserve legacy ID ordering.
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: Optional server filter; changing filters invalidates the cursor. Unfiltered requests preserve legacy ID ordering.
          schema:
            type: string
        - name: owner
          in: query
          required: false
          description: Optional server filter; changing filters invalidates the cursor. Unfiltered requests preserve legacy ID ordering.
          schema:
            type: string
        - name: sort
          in: query
          required: false
          description: Optional server filter; changing filters invalidates the cursor. Unfiltered requests preserve legacy ID ordering.
          schema:
            type: string
            enum:
              - id
              - title
              - updated
              - -updated
        - name: archived
          in: query
          required: false
          description: Optional server filter; changing filters invalidates the cursor. Unfiltered requests preserve legacy ID ordering.
          schema:
            type: string
      responses:
        "200":
          description: Successful operation. Human ownership and resource scopes are validated by the server.
          headers:
            ETag:
              description: Quoted resource version when returning a versioned entity.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: ./schemas/domain.schema.json#/$defs/TaskPage
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
      x-accp-milestone: M1
      x-accp-required-scope: tasks:read
  /tasks/{id}:
    get:
      operationId: getTask
      tags:
        - Tasks
      summary: Read task state
      description: Authoritative state; events are notifications, not a replacement for this read. M2 also allows a scoped Agent Session in the same project.
      x-accp-actors:
        - HUMAN
        - AGENT
      security:
        - HumanBearer: []
        - AgentBearer: []
      parameters:
        - name: id
          in: path
          required: true
          description: Opaque resource identifier; scope is verified server-side.
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Successful operation. Human ownership and resource scopes are validated by the server.
          headers:
            ETag:
              description: Quoted resource version when returning a versioned entity.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: ./schemas/domain.schema.json#/$defs/Task
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
      x-accp-milestone: M1
      x-accp-required-scope: tasks:read
  /tasks/{id}/commands:
    post:
      operationId: commandTask
      tags:
        - Tasks
      summary: Submit, retry or cancel a task
      description: Human Owner or ADMIN submits, retries or cancels execution with the current Task version. RETRY creates no Run until the next authorized claim. Cancellation closes the active Run; DONE is immutable.
      x-accp-actors:
        - HUMAN
      security:
        - HumanBearer: []
      parameters:
        - name: id
          in: path
          required: true
          description: Opaque resource identifier; scope is verified server-side.
          schema:
            type: string
            minLength: 1
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "200":
          description: Successful operation. Human ownership and resource scopes are validated by the server.
          headers:
            ETag:
              description: Quoted resource version when returning a versioned entity.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: ./schemas/domain.schema.json#/$defs/Task
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "412":
          $ref: "#/components/responses/Problem"
        "428":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas/domain.schema.json#/$defs/TaskCommand
      x-accp-milestone: M1
...
  /contexts/{id}/versions:
    post:
      operationId: createContextVersion
      tags:
        - Contexts
      summary: Submit immutable candidate content
      description: "Verify allowed content URI and digest. New version is CANDIDATE even when submitted by an Agent. M1 requires a same-project urn:accp:content: URI from uploadContextContent; digest and media type must match stored bytes. Creation increments the parent Context version."
      x-accp-actors:
        - HUMAN
      security:
        - HumanBearer: []
      parameters:
        - name: id
          in: path
          required: true
          description: Opaque resource identifier; scope is verified server-side.
          schema:
            type: string
            minLength: 1
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "201":
          description: Successful operation. Human ownership and resource scopes are validated by the server.
          headers:
            ETag:
              description: Quoted resource version when returning a versioned entity.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: ./schemas/domain.schema.json#/$defs/ContextVersion
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "412":
          $ref: "#/components/responses/Problem"
        "428":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas/domain.schema.json#/$defs/CreateContextVersion
      x-accp-milestone: M1
    get:
      operationId: listContextVersions
      tags:
        - Contexts
      summary: List immutable Context versions
      description: List immutable Context versions M2 also allows a scoped Agent Session in the same project.
      x-accp-milestone: M1
      x-accp-actors:
        - HUMAN
        - AGENT
      security:
        - HumanBearer: []
        - AgentBearer: []
      parameters:
        - name: id
          in: path
          required: true
          description: Opaque resource ID; scope is verified by the server.
          schema:
            type: string
            minLength: 1
        - &a4
          name: cursor
          in: query
          description: Opaque continuation bound to this collection.
          schema:
            type: string
        - &a5
          name: limit
          in: query
          description: Maximum page size.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Successful M1 operation.
          headers:
            ETag:
              description: Quoted version for versioned resources.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: ./schemas/m1.schema.json#/$defs/ContextVersionPage
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
      x-accp-required-scope: context:read
...
  /me:
    get:
      operationId: getCurrentHuman
      tags:
        - Identity
      summary: Read the provisioned human identity
      description: Read the provisioned human identity
      x-accp-milestone: M1
      x-accp-actors:
        - HUMAN
      security:
        - HumanBearer: []
      parameters: []
      responses:
        "200":
          description: Successful M1 operation.
          headers:
            ETag:
              description: Quoted version for versioned resources.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: ./schemas/m1.schema.json#/$defs/Human
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    HumanBearer:
      type: http
      scheme: bearer
      bearerFormat: Human access token
      description: Validated OIDC-backed human identity for ACCP; project roles checked server-side.
    AgentBearer:
      type: http
      scheme: bearer
      bearerFormat: Scoped ACCP session token
      description: Short-lived audience-bound token; session id alone is not a credential.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique per actor, route and resource; same key with different input returns 409.
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Quoted resource version returned as ETag; 412 on stale version, 428 when absent.
      schema:
        type: string
        pattern: ^"[1-9][0-9]*"$
...
  responses:
    Problem:
      description: Authentication, authorization, version or domain failure; see code and trace_id.
      content:
        application/problem+json:
          schema:
            $ref: ./schemas/domain.schema.json#/$defs/Problem